+++ src/platforms/abstract/abstractParser.js
@@ -1,4 +1,4 @@
-const { NullAttributeError } = require('../../errors/errors')
+const { NullAttributeError } = require('../../error')
 
 class AbstractParser {
     static ID_SEPARATOR = "_"
