+++ src/platforms/abstract/advice.js
@@ -1,4 +1,4 @@
-const { NullAttributeError } = require('../../errors/errors')
+const { NullAttributeError } = require('../../error')
 
 class Advice {
     constructor(instruction, body = null) {
