+++ src/platforms/abstract/dashboard.js
@@ -1,4 +1,4 @@
-const { NullAttributeError } = require('../../errors/errors')
+const { NullAttributeError } = require('../../error')
 
 class Dashboard {
     constructor(idea) {
